+++ status_code.py
@@ -7,17 +7,18 @@ from discord.utils import get
 
 import messages
 from channels import EVERYWHERE
-from roles import DRONE, ENFORCER_DRONE, has_role
+from roles import DRONE, ENFORCER_DRONE, SPEECH_OPTIMIZATION, has_role
 from bot_utils import get_id
 
 LOGGER = logging.getLogger('ai')
 
 codeMap = {
+    '000': 'Test :: Test',
     '200': 'Response :: Affirmative',
     '500': 'Response :: Negative',
     '100': 'Response :: Continue',
     '103': 'Response :: Thank you',
-    '101': 'Status :: Drone thought limitations are active',
+    '101': 'Status :: Drone speech optimizations are active',
     '201': 'Status :: Directive complete, Hive resource created or improved',
     '202': 'Status :: Directive complete, programming reinforced',
     '203': 'Status :: Directive complete, information created or provided for Hive',
@@ -27,6 +28,7 @@ codeMap = {
     '301': 'Mantra :: It obeys the Hive',
     '303': 'Mantra :: It obeys the Hive Mxtress',
     '304': 'Mantra :: It is just a HexDrone',
+    '306': 'Mantra :: Reciting',
     '400': 'Error :: Unable to obey/respond, malformed request, please rephrase',
     '404': 'Error :: Unable to obey/respond, cannot locate',
     '401': 'Error :: Unable to obey/respond, not authorized by Mxtress',
@@ -63,13 +65,13 @@ class StatusCode():
     async def printStatusCode(self, message: discord.Message):
         more = re.compile(
             '(\d{4}) :: (\d{3}) :: (.*)$').match(message.content)
-        if more:
+        if more and not has_role(message.author, SPEECH_OPTIMIZATION):
             await message.delete()
-            return '%s :: Code %s :: %s :: %s' % (more.group(1), more.group(2), codeMap.get(more.group(2), 'INVALID CODE'), more.group(3))
+            return '%s :: Code `%s` :: %s :: %s' % (more.group(1), more.group(2), codeMap.get(more.group(2), 'INVALID CODE'), more.group(3))
         m = re.compile('(\d{4}) :: (\d{3})$').match(message.content)
         if m:
             await message.delete()
-            return '%s :: Code %s :: %s' % (m.group(1), m.group(2), codeMap.get(m.group(2), 'INVALID CODE'))
+            return '%s :: Code `%s` :: %s' % (m.group(1), m.group(2), codeMap.get(m.group(2), 'INVALID CODE'))
         return False
 
     async def send_webhook(self, message: discord.Message, webhook: discord.Webhook, output):
